Support ImageBrush Viewbox and rotation/shearing matrices
This modifies ImageBrush handling to support clipping based on the
Viewbox and multiplies the transform matrix to support rotation or
shearing transformations.

Viewbox is given in units of 1/96 of an inch, which means it depends
on the image resolution.  This modifies image loading (including using
libpng instead of cairo) in order to capture the resolution and wraps
it in a GXPSImage struct.

https://bugzilla.gnome.org/show_bug.cgi?id=751357