Add support for resource dictionaries
It parses the resources and keeps a cache in the form of xml.
Once we are parsing the real xml if we find a resource we
get the cached xml resource and we parse it to properly handle it.

For now only Path Data resources are supported.

Based in a patch from Jason Crain.

https://bugzilla.gnome.org/show_bug.cgi?id=777731