xpstopdf: make the generated pdf to be 96 dpi
The page size we get from the xps metadata is in points. When we
pass this size to cairo we should convert it to 96 dpi since it
is the dpi we have by default on Windows, without this conversion
the generated pdf will be using a bigger page size.
Moreover evince does this conversion as well by itself.

https://bugzilla.gnome.org/show_bug.cgi?id=794248